Reorg and pivoting to constructor-based persistence
After some digging and experimentation by @mrb24, it looks like there is not
presently a practical, sustainable way to serialize a LogicalPlan 🙍.  We had
a quick chat, and the most realistic way forward is to instead preserve the
steps that Mimir took to create a dataframe.  In other words, any API call that
creates a new named DataFrame (CreateView, LoadDataset, CreateLens, etc...) will
now register a "Constructor" with the catalog.

A constructor is really trivial: It takes a set of dependencies (dataframe/name
pairs) and the spark context and returns a dataframe.  There's a bit of
messiness involved in rehydrating the constructor (and I still think it can be
made cleaner), but with a little bit of reflection magic, that part seems to
be working properly.

This update also reorganizes the codebase a little bit to remove 'api.tasks'.
Instead, each request now gets its own file in 'api.requests' containing the
request, the response, and any support goo required.