Enabling Geocoding
Geocoding works in two steps: The training process geocodes all of the addresses
and caches them by staging the address -> lat/long pairs.  When the lens
dataframe is created, we do a left-outer join with the cache table to recover
all of the geocoded addresses.  Note the assumption that the source dataframe
is unchanged from its original form.

The infrastructure is set up to allow for multiple possible geocodings, although
at the moment this facility is not used.

Additionally...
- Catalog now has a "stage and put" convenience method
- The MimirAPI now has configuration parameters to initialize the relevant
  geocoders