Naive Vizual evaluation (re-)implemented and functional.
I've reverted Vizual to the flavor found in the Vizier frontend API.  As part of this change, to
shorten dev times, I've temporarily set aside the vizual repository.  The API no longer relies on
external support to implement Vizual.

The simplified Vizual implemented here should be JSON compatible with Vizier's Vizual cell
description objects.  In other words it should be possible to simply pass a Vizual command JSON
object to Mimir (or more precisely a sequence of operations) wrapped in a VizualRequest and
everything should Just Work™.

The current implementation is about as good as it can get, bearing in mind that scripts presently
consist of individual instructions.  We need to get this deployed to unbreak certain Vizual
operations on the new Mimir (c.f. https://github.com/VizierDB/web-ui/issues/204), but the idea is
that the next steps will be to start migrating to a more beefy version of Vizual where scripts
consist of multiple instructions, and we have proper Spreadsheet-flavored semantics
(c.f., https://github.com/VizierDB/web-ui/issues/7)