Mimir API changes to support properties (re https://github.com/VizierDB/web-ui/issues/236)
The Mimir API catalog now stores an additional Properties field --- Json data set by the user that can be later retrieved by the `query/table` endpoint.  Eventually, we'll want to add some tooling to allow properties to be automatically propagated / derived, but for now this should be good enough to support data profiling.