Adding support for error-aware CSV loading in Mimir v2
With this commit, Mimir replaces the `csv` data loader with a combination of the `text`
data loader and the `from_csv` function.  This makes it possible to record potential errors
that occur during CSV parsing (still assumes that lines are delimited properly, but...).  This
is similar to an analogous feature in Mimir v1, but does not require its own data loader.