Fix ExplainSpec and other misc tweaks for geocoding
The big problem is that in *some* cases, spark's optimizer can push a filter down below the mimir construct that builds ROWIDs.  This is... problematic, since it shifts the ROWID for all records past the first record that fails a filter.  This patch (among other things) changes things so that ROWIDs are hard-assigned to any parquet file created by mimir.

- Fixed parsing String Keys for Caveats in JSON format
- Catalog's stageAndPut now adds rowids if staging requires an intermediate dataframe (fixes ExplainSpec)
- Geocoding lens now only annotates records where the coordinates are unknown
- typo in the loggerutils package name
- Test cases now use simple prettyprinting (works w/ Bloop and SBT)
- Slightly extended ExplainSpec